backend/requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
pillow
opencv-python
python-multipart
scikit-learn
matplotlib
kagglehub
pandas
